Perform monkey-patching after establishing connections
As of ActiveRecord 5.0 monkey-patchings initiated by
ActiveUUID::Patches.apply! do not modify Mysql2Adapter and/or
SQLite3Adapter of ActiveRecord because they are not loaded by the time
the apply! method is invoked.

To load patches properly on loading time, we override
ActiveRecord::Base.establish_connection method to load the patches
immediately after connections are established.